+++ Widgets/SimpleEntry.swift
@@ -1,0 +1,6 @@
+import WidgetKit
+
+struct SimpleEntry: TimelineEntry {
+    let date: Date
+    let configuration: ConfigurationAppIntent
+}
